# apps/api/docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Ownpace — managed API
  version: 0.1.0-rc.1
  summary: Migrate mail, calendar, contacts and files off US cloud, on your own schedule.
  description: |
    The multi-tenant managed edition. Every request is scoped to one tenant by
    the bearer token; the database enforces the same boundary again with
    row-level security, so a token for tenant A cannot read tenant B even if a
    route forgets to filter.

    **Two ideas run through the whole surface and explain most of its shape:**

    - *Nothing consequential happens without a person.* Deleting on the target,
      finishing a migration and moving billing money are gated behind explicit
      evidence, an owner/admin role, or both. Endpoints that could destroy data
      refuse by default and say what is missing.
    - *A failure is reported as itself.* Routes return the reason verbatim
      rather than an empty success. "I found nothing" and "I could not look"
      are different answers and never share a status code.

    Long-running work (sync, cutover, verification, applying a deletion) is
    enqueued, not performed in the request: those endpoints answer `202` with
    an identifier, and a matching read endpoint polls the outcome.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Ownpace
    url: https://github.com/Robbes/Ownpace
servers:
  - url: https://{host}
...
    variables:
      host:
        default: localhost:3001
...
security:
  - bearerAuth: []
...
paths:
...
  /api/tenants:
    get:
      tags: [Tenants]
      summary: List the tenants this token may see
      responses:
        '200':
          description: The caller's tenants.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Tenants]
      summary: Create a tenant — NOT AVAILABLE HERE
      description: |
        Always `501`. Tenant creation is a cross-tenant bootstrap operation and
        cannot run through a tenant-scoped request: the RLS insert policy
        requires the new row's id to equal the current tenant, which a freshly
        created tenant never satisfies. Documented rather than hidden, because
        the honest refusal is the contract — tenants are provisioned by the
        onboarding/seed path.
      responses:
        '501':
          description: Not implemented, with the reason and the path that does work.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/tenants/{tenantId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    get:
      tags: [Tenants]
      summary: Get one tenant
      responses:
        '200':
          description: The tenant.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Tenants]
      summary: Update a tenant's name or settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                settings: { type: object }
      responses:
        '200':
          description: The updated tenant.
          content:
            application/json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Tenants]
      summary: Retired — close the account instead
      description: |
        Gone since workplan 0085. This did a hard delete that cascaded
        twenty-five tables, `invoice` and `audit_log` among them, behind one
        call with no confirmation and no way back — destroying billing records
        that tax law requires be kept.

        It answers 410 naming the replacement rather than 404, so a caller that
        has not been updated is told where to go instead of seeing what looks
        like a routing bug.
      responses:
        '410':
          description: Retired. Use POST /api/tenants/{tenantId}/close.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/tenants/{tenantId}/members:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    get:
      tags: [Tenants]
      summary: List members
      responses:
        '200':
          description: The tenant's members.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Tenants]
      summary: Invite a member
      description: |
        A second invitation for an address that already has a live row is
        refused with `409` naming that row, rather than quietly creating a
        duplicate whose role nobody can predict.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '201':
          description: The invited member.
          content:
            application/json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: An invited or active row for this address already exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/tenants/{tenantId}/members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - name: memberId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Tenants]
      summary: Get one member
      responses:
        '200':
          description: The member.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Tenants]
      summary: Change a member's role
      description: |
        Owner/admin only. Removing your own last owner role is refused — the
        last owner cannot demote themselves out of the tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: The updated member.
          content:
            application/json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Refused — this change would leave the tenant without an owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Tenants]
      summary: Remove a member
      responses:
        '200': { description: Removed. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Refused — removing this member would leave the tenant without an owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A tenant-scoped JWT. Local deployments sign it with `JWT_SECRET`;
        managed SSO validates against a JWKS instead.

        The token's subject must be a MEMBER of the tenant it names — an
        arbitrary subject with a valid signature is refused. Roles (`owner`,
        `admin`, `member`, `viewer`) gate the routes marked owner/admin above,
        for reads as well as writes on billing.
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The request was refused, with the reason.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid, or non-member token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated, but this role may not do this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource in this tenant.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      description: Every refusal has this shape. `message` is meant to be shown to a person as-is.
      properties:
        error: { type: string, examples: ['Conflict'] }
        message:
          type: string
          examples: ['Mapping is paused — review the discovery counts and start it first (POST /start).']
      required: [error, message]
...
    Role:
      type: string
      enum: [owner, admin, member, viewer]
...